+++ aether-core/src/main/java/com/xiaobai1226/aether/core/controller/StaticResourcesController.java
@@ -2,7 +2,7 @@ package com.xiaobai1226.aether.core.controller;
 
 import org.noear.solon.annotation.Component;
 import org.noear.solon.annotation.Mapping;
-import org.noear.solon.core.handle.ModelAndView;
+import org.noear.solon.core.handle.Context;
 
 import static com.xiaobai1226.aether.common.constant.GateWayTagConsts.STATIC_RESOURCES;
 
@@ -10,8 +10,9 @@ import static com.xiaobai1226.aether.common.constant.GateWayTagConsts.STATIC_RES
 public class StaticResourcesController {
 
     @Mapping
-    public ModelAndView base() {
-        ModelAndView model = new ModelAndView("index.html");
-        return model;
+    public void base(Context ctx) {
+        // ModelAndView model = new ModelAndView("index.html");
+        // return model;
+        ctx.forward("/index.html");
     }
 }